requirements.txt
django==2.2.14
django-informixdb==1.9.1
django-admin-honeypot==1.1.0
django-bootstrap4==2.2.0
django-localflavor==3.0
django-loginas==0.3.8
django-taggit==1.3.0
bootstrap_admin==0.4.4
mysqlclient==2.0.1
pycryptodome==3.9.8
pyodbc==4.0.30
python-ldap==3.3.1
python-memcached==1.59
pysftp==0.2.9
wemake-python-styleguide==0.13.3
